+++ src/app/interseptors/http-interceptor/auth-interceptor.ts
@@ -1,8 +1,8 @@
-import { Injectable } from "@angular/core";
-import { HttpRequest, HttpHandler, HttpEvent, HttpInterceptor } from "@angular/common/http";
+import { Injectable } from '@angular/core';
+import { HttpRequest, HttpHandler, HttpEvent, HttpInterceptor } from '@angular/common/http';
 
-import {AuthService} from "../services/auth.service";
-import {Observable} from "rxjs/internal/Observable";
+import {AuthService} from '../../services/auth.service';
+import {Observable} from 'rxjs/internal/Observable';
 
 
 @Injectable()
